cycle_length is a whole number

+++ src/bezier.cpp
@@ -7,7 +7,7 @@ using namespace std;
 using namespace demoloop;
 
 
-const float CYCLE_LENGTH = 10;
+const uint32_t CYCLE_LENGTH = 10;
 
 glm::mat4 bezier_mat = {{-1.f, 3.f, -3.f, 1.f}, {3.f, -6.f, 3.f, 0.f}, {-3.f, 3.f, 0.f, 0.f}, {1.f, 0.f, 0.f, 0.f}};
 glm::vec2 evaluate(const array<glm::vec2, 4> &controlPoints, float t) {
