+++ src/loop007.cpp
@@ -10,7 +10,7 @@ using namespace demoloop;
 #define MAX_VERTS 6
 
 
-const float CYCLE_LENGTH = 6;
+const uint32_t CYCLE_LENGTH = 6;
 
 class Loop7 : public Demoloop {
 public:
