+++ src/loop011.cpp
@@ -8,7 +8,7 @@ using namespace demoloop;
 
 static const uint16_t NUM_VERTS = 60;
 
-const float CYCLE_LENGTH = 3;
+const uint32_t CYCLE_LENGTH = 3;
 static const float RADIUS = 0.3;
 
 class Loop11 : public Demoloop {
