+++ src/loop012.cpp
@@ -9,7 +9,7 @@ using namespace demoloop;
 #define MAX_VERTS 6
 
 
-const float CYCLE_LENGTH = 6;
+const uint32_t CYCLE_LENGTH = 6;
 
 class Loop12 : public Demoloop {
 public:
