+++ src/loop013.cpp
@@ -9,7 +9,7 @@ using namespace demoloop;
 
 static const uint16_t NUM_VERTS = 60;
 
-const float CYCLE_LENGTH = 8;
+const uint32_t CYCLE_LENGTH = 8;
 static const float RADIUS = 0.3;
 
 class Loop13 : public Demoloop {
