+++ src/loop018.cpp
@@ -10,7 +10,7 @@ using namespace demoloop;
 #define NUM_VERTS 12
 
 
-const float CYCLE_LENGTH = 6;
+const uint32_t CYCLE_LENGTH = 6;
 
 class Loop015 : public Demoloop {
 public:
