+++ src/loop023.cpp
@@ -10,7 +10,7 @@ using namespace demoloop;
 #define NUM_VERTS 3
 
 
-const float CYCLE_LENGTH = 5;
+const uint32_t CYCLE_LENGTH = 5;
 
 class Loop021 : public Demoloop {
 public:
