+++ src/loop048.cpp
@@ -7,7 +7,7 @@ using namespace std;
 using namespace demoloop;
 
 
-const float CYCLE_LENGTH = 10;
+const uint32_t CYCLE_LENGTH = 10;
 const uint32_t arms = 7;
 const uint32_t trisPerArm = 50;
 const uint32_t numTris = arms * trisPerArm;
