+++ src/loop059.cpp
@@ -52,7 +52,7 @@ vec4 effect(vec4 color, Image texture, vec2 tc, vec2 screen_coords) {
 #endif
 )===";
 
-const float CYCLE_LENGTH = 7;
+const uint32_t CYCLE_LENGTH = 7;
 class Loop050 : public Demoloop {
 public:
   Loop050() : Demoloop(CYCLE_LENGTH, 720, 720, 0, 0, 0), RADIUS(height / 5), shader({shaderCode, shaderCode}) {
