+++ src/loop060.cpp
@@ -6,7 +6,7 @@ using namespace std;
 using namespace demoloop;
 
 
-const float CYCLE_LENGTH = 5;
+const uint32_t CYCLE_LENGTH = 5;
 
 const static std::string shaderCode = R"===(
 uniform mediump float cycle_ratio;
