+++ src/loop072.cpp
@@ -79,7 +79,7 @@ const uint32_t Z_DUPS = 4;
 const uint32_t SLICES = 25 * Z_DUPS;
 const uint32_t STACKS = 15;
 
-const float CYCLE_LENGTH = 4;
+const uint32_t CYCLE_LENGTH = 4;
 const float FAR_PLANE = 400.0 * Z_DUPS;
 
 class Loop072 : public Demoloop {
