+++ src/pilotpriest_lost_boy.cpp
@@ -8,7 +8,7 @@ using namespace std;
 using namespace demoloop;
 
 
-const float CYCLE_LENGTH = 3;
+const uint32_t CYCLE_LENGTH = 3;
 
 template<uint32_t N>
 typename std::enable_if<N >= 3, array<Vertex, N>>::type
