+++ src/profile_pic.cpp
@@ -6,7 +6,7 @@ using namespace std;
 using namespace demoloop;
 
 
-const float CYCLE_LENGTH = 10;
+const uint32_t CYCLE_LENGTH = 10;
 const uint32_t arms = 5;
 const uint32_t trisPerArm = 20;
 const uint32_t numTris = arms * trisPerArm;
